requirements.txt
asgiref==3.8.1
asttokens==3.0.0
chardet==5.2.0
cssselect2==0.8.0
decorator==5.2.1
Django==5.1.7
django-classy-tags==4.1.0
django-cms==4.1.4
django-extensions==3.2.3
django-extensions-models==1.0
django-filer==3.3.1
django-formtools==2.5.1
django-modelcluster==6.4
django-polymorphic==3.1.0
django-sekizai==4.1.0
django-taggit==6.1.0
django-treebeard==4.7.1
djangocms-admin-style==3.3.1
djangocms-text-ckeditor==5.1.6
djangorestframework==3.15.2
easy-thumbnails==2.10
exceptiongroup==1.2.2
executing==2.1.0
html5lib==1.1
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
lxml==5.3.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.1.0
pip==25.0.1
prompt_toolkit==3.0.50
psycopg2==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
python-decouple==3.8
pytz==2025.1
reportlab==4.3.1
setuptools==75.8.2
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
svglib==1.5.1
tinycss2==1.4.0
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
